Add @react-three/fiber dependency and implement basic scene in Home page
A new dependency, @react-three/fiber, has been added, and it is used to implement a basic 3D scene in the Home page. To handle 3D object loading, a new Loader component has also been created and used for suspension fallback inside the Canvas module of Home page. This commit lays the groundwork for further 3D implementation.