Add Sky component to HomePage
The Sky model has been implemented into the project and added to the Home page. This comes with the addition of a new file, Sky.tsx, where the 3D sky model is imported and managed. The Home page has been updated to include the Sky component in the 3D scene.